[rename🚚] community도메인에서 post 도메인 분리(#21)
