--- CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/community/repository/PostRepository.java
@@ -1,9 +1,0 @@
-package com.culturespot.culturespotdomain.core.community.repository;
-
-import com.culturespot.culturespotdomain.core.community.entity.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-}
