+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/image/dto/ImageDetailRequest.java
