+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/image/dto/ImageDetailResponse.java
