+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/image/entity/Image.java
@@ -1,5 +1,6 @@
 package com.culturespot.culturespotdomain.core.community.entity;
 
+import com.culturespot.culturespotdomain.core.post.entity.Post;
 import jakarta.persistence.*;
 import lombok.*;
