+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/image/repository/ImageRepository.java
