+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/like/entity/Like.java
@@ -1,23 +1,21 @@
 package com.culturespot.culturespotdomain.core.community.entity;
 
+import com.culturespot.culturespotdomain.core.post.entity.Post;
 import com.culturespot.culturespotdomain.core.user.entity.User;
 import jakarta.persistence.*;
 
 @Entity
-@Table(name = "comments")
-public class Comment {
+@Table(name = "likes")
+public class Like {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private Long likeId;
 
     @ManyToOne
     @JoinColumn(name = "userId", nullable = false)
     private User user;
 
     @ManyToOne
-    @JoinColumn(name = "post_id", nullable = false)
+    @JoinColumn(name = "postId")
     private Post post;
-
-    @Column
-    private String content;
 }
