+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/like/repository/LikeRepository.java
